[WOOR-132] refactor: resolver 패키지 분리
