+++ src/main/java/com/musseukpeople/woorimap/auth/presentation/resolver/Tokens.java
@@ -1,4 +1,4 @@
-package com.musseukpeople.woorimap.auth.presentation.dto;
+package com.musseukpeople.woorimap.auth.presentation.resolver;
 
 import lombok.Getter;
